requirements.txt
asgiref==3.3.4
Django
djangorestframework
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary>=2.8
django-filter
dj.choices
djangorestframework-gis
django-rest-swagger
django-extra-fields
drf_extra_fields